cloud_functions/document_classifier/requirements.txt
google-cloud-documentai==1.4.2
google-cloud-storage==2.4.0
google-cloud-pubsub==2.13.3
google-cloud-firestore==2.5.3